scripts/spread_pwc/requirements.txt
kafka-python==2.0.2
cx_Oracle==8.1.0
pyyaml==5.4.1
